Changed the Node Tree ID from a long to int. Also added a new operator for getting the node id as a combined long from node tree and local id.
git-svn-id: https://svn.apache.org/repos/asf/incubator/vxquery@1468304 13f79535-47bb-0310-9956-ffa450edef68

+++ trunk/vxquery/vxquery-core/src/main/java/org/apache/vxquery/xmlparser/ITreeNodeIdProvider.java
@@ -17,5 +17,5 @@
 package org.apache.vxquery.xmlparser;
 
 public interface ITreeNodeIdProvider {
-    public long getId();
+    public int getId();
 }
